pyproject.toml
[project]
name = "sprint-summary-slack-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jira>=3.10.5",
    "python-dotenv>=1.1.1",
    "slack-bolt>=1.23.0",
    "slack-sdk>=3.36.0",
    "mcp>=0.1.0",
    "google-genai>=1.32.0",
    "fastmcp>=2.11.3",
]
